Refactor app with new directory structure
The most evident is definitely the change in the directory structure.
But there were other important fixes in this commit. Especially the
addition of negative integer support. The other major changes being
mostly in the subtract and divide functions, which are now closely
modeled to how they're defined in Clojure