Update TODOs
Some items were checked off and new items are now a part of this
ever-growing list of TODOs. It is so amazing that I see some problem
every time with the current codebase and that is noted down as a TODO.

Perfection is not what I'm after but instead to try and stay as close
to something familiar writing code in Clojure.

Loving the fact that my commit messages are also places for my notes.
Since Github doesn't let me push these notes.