Add support to generate executable binary for REPL
This is a minor update using which you can generate an executable
binary of the REPL. This is possible using a tool named nexe. You
could use the Makefile to generate the binary and install it on
your system.

There was a minor update to the REPL as well. In this update, I
changed the way responses were being output.